Move Arr tests.
